fix(messenger): smooth navigation
