+++ js/packages/components/NotificationProvider.tsx
@@ -1,29 +1,32 @@
 import React from 'react'
-import { InAppNotificationProvider, withInAppNotification } from 'react-native-in-app-notification'
+import {
+	InAppNotificationProvider,
+	// withInAppNotification
+} from 'react-native-in-app-notification'
 
-import { useMsgrContext } from '@berty-tech/store/hooks'
+// import { useMsgrContext } from '@berty-tech/store/hooks'
 
 import NotificationBody from './NotificationBody'
 
-const NotificationBridge: React.FC = withInAppNotification(({ showNotification }: any) => {
-	const { addNotificationListener, removeNotificationListener } = useMsgrContext()
+// const NotificationBridge: React.FC = withInAppNotification(({ showNotification }: any) => {
+// 	const { addNotificationListener, removeNotificationListener } = useMsgrContext()
 
-	React.useEffect(() => {
-		const listener = (evt: any) => {
-			showNotification({
-				title: evt.payload.title,
-				message: evt.payload.body,
-				additionalProps: evt,
-			})
-		}
-		addNotificationListener(listener)
+// 	React.useEffect(() => {
+// 		const listener = (evt: any) => {
+// 			showNotification({
+// 				title: evt.payload.title,
+// 				message: evt.payload.body,
+// 				additionalProps: evt,
+// 			})
+// 		}
+// 		addNotificationListener(listener)
 
-		return () => {
-			removeNotificationListener(listener)
-		}
-	}, [showNotification, addNotificationListener, removeNotificationListener])
-	return null
-})
+// 		return () => {
+// 			removeNotificationListener(listener)
+// 		}
+// 	}, [showNotification, addNotificationListener, removeNotificationListener])
+// 	return null
+// })
 
 const NotificationProvider: React.FC = ({ children }) => (
 	<InAppNotificationProvider
@@ -32,7 +35,7 @@ const NotificationProvider: React.FC = ({ children }) => (
 		closeInterval={5000}
 	>
 		<>
-			<NotificationBridge />
+			{/* <NotificationBridge /> */}
 			{children}
 		</>
 	</InAppNotificationProvider>
