+++ js/packages/components/settings/Mode.tsx
@@ -14,6 +14,10 @@ import { HeaderSettings } from '../shared-components/Header'
 import { ButtonSetting } from '../shared-components/SettingsButtons'
 import { useNavigation as useReactNavigation } from '@react-navigation/native'
 import { SwipeNavRecognizer } from '../shared-components/SwipeNavRecognizer'
+import i18n from '@berty-tech/berty-i18n'
+import { DropDownPicker } from '@berty-tech/components/shared-components/DropDownPicker'
+import { languages } from '@berty-tech/berty-i18n/locale/languages'
+import { PersistentOptionsKeys } from '@berty-tech/store/context'
 
 //
 // Mode
@@ -42,110 +46,29 @@ const BodyMode: React.FC<BodyModeProps> = () => {
 	)
 
 	const ctx = useMsgrContext()
+	const items: any = Object.entries(languages).map(([key, attrs]) => ({
+		label: attrs.localName,
+		value: key,
+	}))
 
-	// const isPrefMode = ctx.persistentOptions.preset.value === 'performance'
-	// const enableNotif = ctx.persistentOptions.notifications.enable
-
-	// const items: any = Object.entries(languages).map(([key, attrs]) => ({
-	// 	label: attrs.localName,
-	// 	value: key,
-	// }))
-	//
-	// items.push({ label: 'Debug', value: 'cimode' })
-
+	items.push({ label: 'Debug', value: 'cimode' })
 	return (
 		<Translation>
 			{(t: any): React.ReactNode => (
 				<View style={[flex.tiny, padding.medium, margin.bottom.medium]}>
-					{/*<DropDownPicker*/}
-					{/*	items={items}*/}
-					{/*	defaultValue={ctx.persistentOptions?.i18n.language}*/}
-					{/*	onChangeItem={async (item: any) => {*/}
-					{/*		await ctx.setPersistentOption({*/}
-					{/*			type: PersistentOptionsKeys.I18N,*/}
-					{/*			payload: {*/}
-					{/*				language: item.value,*/}
-					{/*			},*/}
-					{/*		})*/}
-					{/*		await i18n.changeLanguage(item.value)*/}
-					{/*	}}*/}
-					{/*/>*/}
-					{/*<ButtonSetting*/}
-					{/*	name={t('settings.mode.notifications-button.title')}*/}
-					{/*	icon='bell-outline'*/}
-					{/*	iconColor={color.blue}*/}
-					{/*	state={{*/}
-					{/*		value: enableNotif*/}
-					{/*			? t('settings.mode.notifications-button.tag-enabled')*/}
-					{/*			: t('settings.mode.notifications-button.tag-disabled'),*/}
-					{/*		color: enableNotif ? color.green : color.red,*/}
-					{/*		bgColor: enableNotif ? color.light.green : color.light.red,*/}
-					{/*	}}*/}
-					{/*	onPress={() => navigation.navigate('Settings.Notifications')}*/}
-					{/*/>*/}
-					{/*<ButtonSetting*/}
-					{/*	name={t('settings.mode.bluetooth-button.title')}*/}
-					{/*	icon='bluetooth-outline'*/}
-					{/*	iconColor={color.blue}*/}
-					{/*	onPress={() => navigation.navigate('Settings.Bluetooth')}*/}
-					{/*/>*/}
-					{/*<ButtonSetting*/}
-					{/*	name={t('settings.mode.app-mode-button.title')}*/}
-					{/*	icon='options-outline'*/}
-					{/*	iconSize={30}*/}
-					{/*	iconColor={color.blue}*/}
-					{/*	actionIcon='arrow-ios-forward'*/}
-					{/*	state={{*/}
-					{/*		value: isPrefMode*/}
-					{/*			? t('settings.mode.app-mode-button.performance-tag')*/}
-					{/*			: t('settings.mode.app-mode-button.privacy-tag'),*/}
-					{/*		color: color.white,*/}
-					{/*		bgColor: isPrefMode ? color.blue : color.red,*/}
-					{/*		stateIcon: isPrefMode ? 'flash-outline' : 'lock-outline',*/}
-					{/*		stateIconColor: color.white,*/}
-					{/*	}}*/}
-					{/*	disabled*/}
-					{/*>*/}
-					{/*	<Text*/}
-					{/*		style={[*/}
-					{/*			column.item.right,*/}
-					{/*			_styles.buttonListUnderStateText,*/}
-					{/*			isPrefMode ? text.color.blue : text.color.red,*/}
-					{/*			margin.bottom.small,*/}
-					{/*		]}*/}
-					{/*	>*/}
-					{/*		{t('settings.mode.app-mode-button.description-tag')}*/}
-					{/*	</Text>*/}
-					{/*	<View style={[padding.right.small]}>*/}
-					{/*		<ButtonSettingItem*/}
-					{/*			value={t('settings.mode.app-mode-button.first-bullet-point')}*/}
-					{/*			// color='rgba(43,46,77,0.8)'*/}
-					{/*			icon={isPrefMode ? 'checkmark-circle-2' : 'close-circle'}*/}
-					{/*			iconColor={isPrefMode ? color.blue : color.red}*/}
-					{/*			disabled*/}
-					{/*			styleText={[text.color.grey]}*/}
-					{/*			styleContainer={[margin.bottom.tiny]}*/}
-					{/*		/>*/}
-					{/*		<ButtonSettingItem*/}
-					{/*			value={t('settings.mode.app-mode-button.second-bullet-point')}*/}
-					{/*			color='rgba(43,46,77,0.8)'*/}
-					{/*			icon={isPrefMode ? 'checkmark-circle-2' : 'close-circle'}*/}
-					{/*			iconColor={isPrefMode ? color.blue : color.red}*/}
-					{/*			disabled*/}
-					{/*			styleText={[text.color.grey]}*/}
-					{/*			styleContainer={[margin.bottom.tiny]}*/}
-					{/*		/>*/}
-					{/*		<ButtonSettingItem*/}
-					{/*			value={t('settings.mode.app-mode-button.third-bullet-point')}*/}
-					{/*			color='rgba(43,46,77,0.8)'*/}
-					{/*			icon={isPrefMode ? 'checkmark-circle-2' : 'close-circle'}*/}
-					{/*			iconColor={isPrefMode ? color.blue : color.red}*/}
-					{/*			disabled*/}
-					{/*			styleText={[text.color.grey]}*/}
-					{/*			styleContainer={[margin.bottom.tiny]}*/}
-					{/*		/>*/}
-					{/*	</View>*/}
-					{/*</ButtonSetting>*/}
+					<DropDownPicker
+						items={items}
+						defaultValue={ctx.persistentOptions?.i18n.language}
+						onChangeItem={async (item: any) => {
+							await ctx.setPersistentOption({
+								type: PersistentOptionsKeys.I18N,
+								payload: {
+									language: item.value,
+								},
+							})
+							await i18n.changeLanguage(item.value)
+						}}
+					/>
 					<ButtonSetting
 						name={t('settings.home.network-button')}
 						icon='earth'
@@ -153,13 +76,6 @@ const BodyMode: React.FC<BodyModeProps> = () => {
 						iconColor={color.blue}
 						onPress={() => navigation.navigate('Settings.NetworkMap')}
 					/>
-					{/*<ButtonSetting*/}
-					{/*	name={t('settings.mode.dark-mode-button')}*/}
-					{/*	icon='moon-outline'*/}
-					{/*	iconColor={color.blue}*/}
-					{/*	toggled*/}
-					{/*	disabled*/}
-					{/*/>*/}
 					<ButtonSetting
 						name={t('settings.mode.receive-contact-requests-button')}
 						icon='person-done-outline'
