+++ js/packages/store/provider.tsx
@@ -35,10 +35,12 @@ export const MsgrProvider: React.FC<any> = ({ children, daemonAddress, embedded 
 
 	useEffect(() => {
 		initialLaunch(dispatch, embedded)
+		return () => {}
 	}, [embedded])
 
 	useEffect(() => {
 		openingDaemon(dispatch, state.appState, state.selectedAccount)
+		return () => {}
 	}, [embedded, state.appState, state.selectedAccount])
 
 	useEffect(() => openingClients(dispatch, state.appState, eventEmitter, daemonAddress, embedded), [
